Move the person entity to the database module
The core should have as few dependencies as possible. Until now, the class person within the core was annotated with @Entity, so JPA recognizes it as an entity. However, this reverses the actual meaning of Clean Architecture. Adapters should implement / use the core. When annotating with @Entity, for example, a Neo4j graph database cannot be used without problems. The easiest way is to use a mapper, for example the ModelMapper.